refactor: Remove non-alt version
BREAKING CHANGE: This package no longer require `ts-morph` and does not take inn file argument. Please consult our README.md for more information
