+++ package.json
@@ -22,9 +22,9 @@
     "@types/faker": "^5.1.7",
     "@types/jest": "~26.0.21",
     "@types/node": "~14.14.35",
-    "typedoc": "^0.20.1",
+    "typedoc": "^0.20.34",
     "typedoc-plugin-markdown": "^3.0.8",
-    "@typescript-eslint/eslint-plugin": "~4.18.0",
+    "@typescript-eslint/eslint-plugin": "~4.19.0",
     "@typescript-eslint/parser": "~4.19.0",
     "eslint": "~7.22.0",
     "eslint-config-prettier": "~8.1.0",
@@ -41,7 +41,6 @@
   },
   "dependencies": {
     "runtypes": "^5.0.2",
-    "ts-morph": "^10.0.1",
     "typescript": "~4.1.3"
   }
 }
